add id number validation

+++ controllers/warehouse-controller.js
@@ -66,19 +66,27 @@ const getInventories = async (req, res) => {
 
 const remove = async (req, res) => {
   try {
+    const { id } = req.params;
+
+    if (isNaN(id)) {
+      return res.status(400).json({
+        message: `Warehouse ID ${id} is invalid`,
+      });
+    }
+
     const warehouseDeleted = await knex("warehouses")
-      .where({ id: req.params.id })
+      .where({ id: id })
       .delete();
 
     if (warehouseDeleted === 0) {
       return res
         .status(404)
-        .json({ message: `Warehouse with ID ${req.params.id} not found` });
+        .json({ message: `Warehouse with ID ${id} not found` });
     }
     res.sendStatus(204);
   } catch (err) {
     res.status(500).json({
-      message: `Unable to delete warehouse with ID ${req.params.id}`,
+      message: `Unable to delete warehouse with ID ${id}`,
     });
   }
 };
